+++ draw_wins.h
@@ -1,4 +1,4 @@
-void draw_with_linebreaks (struct Win *, char *, int);
+void draw_with_linebreaks (struct Win *, char *, uint16_t);
 void draw_text_from_bottom (struct Win *, char *);
 void draw_log_win (struct Win *);
 void draw_map_win (struct Win *);
